fix bugs R check

+++ src/to_index.cpp
@@ -815,7 +815,7 @@ SEXP cpp_to_index(SEXP x){
       all_vecs.push_back(rvec);
       if(k == 0){
         n = Rf_length(VECTOR_ELT(x, 0));
-      } else if(Rf_length(VECTOR_ELT(x, k)) != n){
+      } else if((size_t) Rf_length(VECTOR_ELT(x, k)) != n){
         Rf_error("All the vectors to turn into an index must be of the same length. This is currently not the case.");
       }
     }
